requirements.txt
opencv-python
atari-py
pyprind 
psutil
tensorboard
black
rope
pylint
pyyaml
typed-argument-parser
gym
debugpy
loguru
wandb
tabulate
moviepy
imageio
progressbar2
matplotlib
nfnets-pytorch
jupyterlab
seaborn
celluloid
xvfbwrapper